+++ Panther/src/ComptimeExecutor.h
@@ -27,13 +27,23 @@ namespace pcit::panther{
 			auto init() -> std::string; // string is error message if initalization fails (empty if successful)
 			auto deinit() -> void;
 
+			EVO_NODISCARD auto isInitialized() const -> bool { return this->data != nullptr; }
+
 			EVO_NODISCARD auto runFunc(
 				const ASG::Func::LinkID& link_id, evo::ArrayProxy<ASG::Expr> params, ASGBuffer& asg_buffer
 			) -> evo::SmallVector<ASG::Expr>;
 
-	
+
+			auto addFunc(const ASG::Func::LinkID& func_link_id) -> void;
+
+		private:
+			auto restart_engine_if_needed() -> void;
+
 		private:
 			class Context& context;
+
+			mutable std::shared_mutex mutex{};
+			bool requires_engine_restart = false;
 
 			struct Data;
 			Data* data = nullptr;